Changed the editing to make it much more intuitive. Abstacted a lote of code from create_trip.dart to decouple the ActionChips.

+++ lib/tiles/marker_tile.dart
@@ -82,6 +82,7 @@ class _MarkerTileState extends State<MarkerTile> {
           ),
           onPressed: () => {},
         ),
+
         children: <Widget>[
           SingleChildScrollView(
             child: Column(
@@ -163,7 +164,7 @@ class _MarkerTileState extends State<MarkerTile> {
   getRating() async {
     getPointOfInterestRating(widget.url).then((ratingMap) {
       setState(() {
-        score = ratingMap['rating'];
+        score = ratingMap['rating'].toDouble();
         scored = ratingMap['scored'];
       });
     });
